Create temperature-multi-ds18b20.md
In this example that works with ConfigurableFirmata (with its high baudrate), we wanted to keep automatic port detection for the board. This is why we are using CustomIO in the Board initialization. Then we call the sendOneWireSearch on Firmata to retreive all the connected DS18B20 ti finaly initialize each sensor with its address.